Fix - create/close issues method reused
