+++ apps/codelab/src/app/codelabs/extra/feedback-page/feedback-page.component.html
@@ -34,7 +34,7 @@
           </div>
           <div>{{ message.comment }}</div>
           <div *ngIf="!message.url">
-            <button (click)="createAnIssue(message)">Take</button>
+            <button (click)="createIssue(message)">Take</button>
             <button (click)="createClosedIssue(message, '[Duplicate]')">
               Close [Duplicate]
             </button>
