+++ apps/codelab/src/app/codelabs/extra/feedback-page/feedback-page.component.ts
@@ -1,7 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { AngularFireDatabase, AngularFireList } from '@angular/fire/database';
-import { AngularFireAuth } from '@angular/fire/auth';
-import * as firebase from 'firebase/app';
 import { Message } from '@codelab/feedback/src/lib/message';
 import { BehaviorSubject, combineLatest, Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
@@ -18,7 +16,7 @@ function groupBy(feedback: Array<Message>, grouping: Grouping) {
     return comment;
   }, {});
 
-  return Object.keys(result).map(key => ({ key, value: result[key] }));
+  return Object.keys(result).map(key => ({key, value: result[key]}));
 }
 
 function normalize(feedback: Array<any>) {
@@ -66,95 +64,12 @@ export class FeedbackPageComponent implements OnInit {
   messages$: Observable<{ key: string; value: Message }[]>;
   filter$ = new BehaviorSubject<Filter>('notDone');
   group$ = new BehaviorSubject<Grouping>('href');
-  githubAuth;
   private feedback$: AngularFireList<any[]>;
 
   constructor(
     private database: AngularFireDatabase,
-    private afAuth: AngularFireAuth,
     private ghService: GithubService
   ) {
-    afAuth.authState.subscribe(authData => {
-      if (authData === null) {
-        this.login();
-      } else {
-        this.githubAuth = authData;
-      }
-    });
-  }
-
-  async login() {
-    const provider = new firebase.auth.GithubAuthProvider().addScope('repo');
-    this.githubAuth = await this.afAuth.auth.signInWithPopup(provider);
-  }
-
-  isDone(message) {
-    this.database
-      .object(`feedback/${message.key}`)
-      .update({ isDone: !message.isDone });
-  }
-
-  generateIssueBody(message) {
-    return `${message.comment}
-Author: ${message.name}
-Slide: [Local](http://localhost:4200${
-      message.href
-    }),[Public](https://angular-presentation.firebaseapp.com${message.href})`;
-  }
-
-  async createAnIssue(message) {
-    if (!this.githubAuth.credential) {
-      await this.login();
-    }
-
-    this.ghService
-      .createIssue(
-        {
-          title: message.comment.substring(0, 150),
-          body: this.generateIssueBody(message)
-        },
-        this.githubAuth.credential.accessToken
-      )
-      .subscribe((responseData: any) => {
-        this.isDone(message);
-        this.database
-          .object(`feedback/${message.key}`)
-          .update({ url: responseData.html_url });
-        window.open(responseData.html_url);
-      });
-  }
-
-  async createClosedIssue(message, reason) {
-    if (!this.githubAuth.credential) {
-      await this.login();
-    }
-
-    this.ghService
-      .createIssue(
-        {
-          title: reason + ' ' + message.comment.substring(0, 150),
-          body: this.generateIssueBody(message)
-        },
-        this.githubAuth.credential.accessToken
-      )
-      .subscribe((responseData: any) => {
-        // Until we get a better UI
-        console.log(responseData.html_url);
-        this.database
-          .object(`feedback/${message.key}`)
-          .update({ url: responseData.html_url });
-        this.ghService
-          .closeIssue(
-            { state: 'closed' },
-            responseData.number,
-            this.githubAuth.credential.accessToken
-          )
-          .subscribe(res => {
-            if (res) {
-              this.isDone(message);
-            }
-          });
-      });
   }
 
   ngOnInit() {
@@ -163,8 +78,14 @@ Slide: [Local](http://localhost:4200${
       this.feedback$.snapshotChanges().pipe(map(normalize)),
       this.filter$
     ).pipe(map(filter));
-    this.messages$ = combineLatest(filteredMessages$, this.group$).pipe(
-      map(group)
-    );
+    this.messages$ = combineLatest(filteredMessages$, this.group$).pipe(map(group));
+  }
+
+  createIssue(message) {
+    this.ghService.createIssue(message);
+  }
+
+  createClosedIssue(message, reason) {
+    this.ghService.createClosedIssue(message, reason);
   }
 }
