+++ libs/feedback/src/lib/feedback-widget/feedback-widget.component.html
@@ -21,7 +21,7 @@
                     <img alt="" src="/assets/images/icons/icon-tools.png"/>
                   </button>
                   <div class="dropdown-content">
-                    <a (click)="createAnIssue(item)"
+                    <a (click)="createIssue(item)"
                        class="dropdown-content-item">
                       Take
                     </a>
